#900 make min/max fields floats 2

+++ server/core/src/main/java/cc/blynk/server/core/model/widgets/outputs/graph/EnhancedHistoryGraph.java
@@ -38,9 +38,9 @@ public class EnhancedHistoryGraph extends Widget {
 
     public boolean overrideYAxis;
 
-    public int yAxisMin;
+    public float yAxisMin;
 
-    public int yAxisMax;
+    public float yAxisMax;
 
     public boolean isPercentMaxMin;
